requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic[email]==2.6.3
pydantic-settings==2.2.1
python-multipart==0.0.9

# Database & Migrations
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# Security
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0

# Performance & Scaling
redis==5.0.1
prometheus-client==0.20.0

# Utilities
httpx==0.27.0
click==8.1.7